+++ .coveralls 
@@ -1,0 +1,2 @@
+service_name: travis-pro
+repo_token: S8RsISTFPftFaeQmZLHCuJH6uvM7sXlm1
