+++ .env
@@ -1,0 +1,1 @@
+export COINMARKETCAP_API_KEY=4e544746-fed4-4385-88b4-a3abed8321a1
