+++ .travis.yml
@@ -1,0 +1,19 @@
+dist: trusty
+sudo: required
+language: node_js
+node_js:
+  - '12'
+cache:
+  directories:
+    - node_modules
+script:
+  - yarn format
+  - yarn lint
+  - yarn test
+  - yarn coverage
+after_success:
+  - cat coverage/lcov.info | npx coveralls
+notifications:
+  email:
+    - dev-support@ampleforth.org
+  slack: fragmentsorg:2bN0L0FCZZ0illBRyPucRZkV
