+++ contracts/_interfaces/IBridgeGateway.sol
