+++ contracts/_interfaces/IXCAmple.sol
