+++ contracts/_interfaces/IXCAmpleController.sol
