+++ contracts/base-chain/bridge-gateways/AMPLChainBridgeGateway.sol
@@ -4,9 +4,9 @@ pragma solidity 0.6.12;
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import "../../bridge-interfaces/IAmpleforthPolicy.sol";
-import "../../bridge-interfaces/ITokenVault.sol";
-import "../../bridge-interfaces/IBridgeGateway.sol";
+import "../../_interfaces/IAmpleforthPolicy.sol";
+import "../../_interfaces/ITokenVault.sol";
+import "../../_interfaces/IBridgeGateway.sol";
 
 /**
  * @title AMPLChainBridgeGateway: AMPL-ChainBridge Gateway Contract
