+++ contracts/satellite-chain/xc-ampleforth/XCAmple.sol
