+++ test/unit/satellite-chain/xc-ampleforth/xc_ample_burn.js
@@ -18,7 +18,7 @@ async function setupContracts () {
 
   // deploy upgradable token
   const factory = await ethers.getContractFactory(
-    'contracts/xc-ampleforth/XCAmple.sol:XCAmple',
+    'contracts/satellite-chain/xc-ampleforth/XCAmple.sol:XCAmple',
   );
   xcAmple = await upgrades.deployProxy(
     factory.connect(deployer),
