+++ test/unit/satellite-chain/xc-ampleforth/xc_controller_bridge.js
@@ -26,7 +26,7 @@ async function setupContracts () {
 
   // deploy upgradable token
   const factory = await ethers.getContractFactory(
-    'contracts/xc-ampleforth/XCAmpleController.sol:XCAmpleController',
+    'contracts/satellite-chain/xc-ampleforth/XCAmpleController.sol:XCAmpleController',
   );
   controller = await upgrades.deployProxy(
     factory.connect(deployer),
